@@ -54,8 +54,8 @@
     "gatsby-transformer-remark": "^2.6.39",
     "node-sass": "^4.13.0",
     "nothing": "0.1.0",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
     "react-helmet": "5.2.0",
     "react-md": "^1.12.1",
     "react-netlify-identity": "^0.1.9",
