@@ -1,38 +1,24 @@
 import React from 'react';
-import { handleLogin, handleLogout, isLoggedIn } from 'services/auth'
+import IdentityModal, { useIdentityContext } from 'react-netlify-identity-widget'
 
-class Login extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            login: isLoggedIn()
-        }
-        this._handleLogin = this._handleLogin.bind(this);
+const Login = () => {
+    const [dialog, setDialog] = React.useState(false);
+    const identity = useIdentityContext();
 
-        this._handleLogout = this._handleLogout.bind(this);
-    }
+    return(
+        <div>
+            <button className="btn" style={{ maxWidth: 400, background: 'darkgreen' }} onClick={() => setDialog(true)}>
+                {
+                    identity && identity.isLoggedIn ? 'LOG OUT' : 'LOG IN'
+                }
+            </button>
 
-    _handleLogin = () => {
-        handleLogin()
-        this.setState({login: isLoggedIn()})
-    }
-
-    _handleLogout = () => {
-        handleLogout()
-        this.setState({login: isLoggedIn()})
-    }
-
-    render() {
-       return(
-           !this.state.login ? 
-            (<button className="" onClick={handleLogin}>
-                Login
-            </button>) :
-            (<button className="" onClick={handleLogout}>
-            Logout
-             </button>)
-        )
-    }
+            <IdentityModal
+                showDialog={dialog}
+                onCloseDialog={() => setDialog(false)}
+            />
+        </div>
+    )
 }
 
 export default Login;
